move athena client from common to worksheets/aws
