--- backend/dataall/modules/common/athena/__init__.py
@@ -1,1 +1,0 @@
-""" Common code for Athena requests """
