+++ backend/dataall/modules/datasets/api/table/resolvers.py
@@ -2,6 +2,7 @@ import json
 import logging
 
 from botocore.exceptions import ClientError
+from pyathena import connect
 
 from dataall import db
 from dataall.modules.datasets.api.dataset.resolvers import get_dataset
@@ -10,7 +11,6 @@ from dataall.aws.handlers.service_handlers import Worker
 from dataall.aws.handlers.sts import SessionHelper
 from dataall.db import models
 from dataall.db.api import ResourcePolicy, Glossary
-from dataall.modules.common.athena.athena_client import run_athena_query
 from dataall.modules.datasets.db.models import DatasetTable, Dataset
 from dataall.modules.datasets.services.dataset_service import DatasetService
 from dataall.modules.datasets.services.dataset_permissions import UPDATE_DATASET_TABLE, PREVIEW_DATASET_TABLE
@@ -109,6 +109,7 @@ def preview(context, source, tableUri: str = None):
         env = db.api.Environment.get_environment_by_uri(session, dataset.environmentUri)
         env_workgroup = {}
         boto3_session = SessionHelper.remote_session(accountid=table.AWSAccountId)
+        creds = boto3_session.get_credentials()
         try:
             env_workgroup = boto3_session.client(
                 'athena', region_name=env.region
@@ -119,15 +120,18 @@ def preview(context, source, tableUri: str = None):
                 f'due to: {e}'
             )
 
-        SQL = f'select * from "{table.GlueDatabaseName}"."{table.GlueTableName}" limit 50'  # nosec
-        cursor = run_athena_query(
-            session=boto3_session,
+        connection = connect(
+            aws_access_key_id=creds.access_key,
+            aws_secret_access_key=creds.secret_key,
+            aws_session_token=creds.token,
             work_group=env_workgroup.get('WorkGroup', {}).get('Name', 'primary'),
             s3_staging_dir=f's3://{env.EnvironmentDefaultBucketName}/preview/{dataset.datasetUri}/{table.tableUri}',
-            region=table.region,
-            sql=SQL
+            region_name=table.region,
         )
+        cursor = connection.cursor()
 
+        SQL = f'select * from "{table.GlueDatabaseName}"."{table.GlueTableName}" limit 50'  # nosec
+        cursor.execute(SQL)
         fields = []
         for f in cursor.description:
             fields.append(json.dumps({'name': f[0]}))
