+++ backend/dataall/modules/worksheets/aws/__init__.py
@@ -1,0 +1,1 @@
+"""Contains code that send requests to AWS Athena """
